Add Installments subtitle to CardComponent when installments are available
