+++ core/src/main/java/com/adyen/checkout/core/common/localization/CheckoutLocalizationKey.kt
@@ -36,6 +36,7 @@ enum class CheckoutLocalizationKey {
     CARD_KCP_CARD_PASSWORD_INVALID,
     CARD_POSTAL_CODE,
     CARD_POSTAL_CODE_INVALID,
+    CARD_INSTALLMENTS,
     CARD_INSTALLMENTS_TITLE,
     CARD_INSTALLMENTS_SUBTITLE,
     CARD_INSTALLMENTS_ONE_TIME,
