+++ core/src/main/res/values/strings.xml
@@ -38,6 +38,7 @@
     <string name="checkout_card_kcp_password_invalid" translatable="false">Invalid password</string>
     <string name="checkout_card_postal_code" translatable="false">Postal code</string>
     <string name="checkout_card_postal_code_invalid" translatable="false">Invalid postal code</string>
+    <string name="checkout_card_installments" translatable="false">Installments</string>
     <string name="checkout_card_installments_title" translatable="false">Number of installments</string>
     <string name="checkout_card_installments_subtitle" translatable="false">Select the number of installments you would like to use</string>
     <string name="checkout_card_installments_option_one_time" translatable="false">One time payment</string>
